add a 'newline' progress mode and a rate-limit for progress messages
This is to make the output more suited for CI. CI environments all
have 'no output' timeouts (and for good reasons). Disabling progress
completely doesn't work well for long builds, and having the progress
just outputs a mess on most console displays.

The new mode will display progress, but with each new progress line
on a single line. Along with the rate-limiting, it allows to know
that things are progressing without having a build log that's unreadable.